+++ certbot/certbot/_internal/tests/crypto_util_test.py
@@ -207,6 +207,14 @@ class MakeKeyTest(unittest.TestCase):
                            match=re.escape('Invalid key_type specified: unf.  Use [rsa|ecdsa]')):
             make_key(2048, key_type='unf')
 
+    def test_for_pkcs8_format(self):
+        from certbot.crypto_util import make_key
+
+        # PKCS#1 format will instead have text like "BEGIN RSA PRIVATE KEY" or "BEGIN EC PRIVATE
+        # KEY"
+        assert b"BEGIN PRIVATE KEY" in make_key(2048)
+        assert b"BEGIN PRIVATE KEY" in make_key(elliptic_curve='secp256r1', key_type='ecdsa')
+
 
 class VerifyCertSetup(unittest.TestCase):
     """Refactoring for verification tests."""
@@ -360,8 +368,8 @@ class GetNamesFromCertTest(unittest.TestCase):
             self._call(test_util.load_vector('cert-5sans_512.pem'))
 
     def test_parse_non_cert(self):
-        with pytest.raises(OpenSSL.crypto.Error):
-            self._call("hello there")
+        with pytest.raises(ValueError):
+            self._call(b"hello there")
 
 
 class GetNamesFromReqTest(unittest.TestCase):