+++ certbot/certbot/crypto_util.py
@@ -8,7 +8,6 @@ import datetime
 import hashlib
 import logging
 import re
-from typing import Callable
 from typing import List
 from typing import Optional
 from typing import Set
@@ -35,7 +34,6 @@ from cryptography.hazmat.primitives.serialization import PrivateFormat
 import josepy
 from OpenSSL import crypto
 from OpenSSL import SSL
-import pyrfc3339
 
 from acme import crypto_util as acme_crypto_util
 from certbot import errors
@@ -172,42 +170,40 @@ def csr_matches_pubkey(csr: bytes, privkey: bytes) -> bool:
     """
     req = x509.load_pem_x509_csr(csr)
     pkey = serialization.load_pem_private_key(privkey, password=None)
-    # This would be better written as `req.public_key() == pkey.public_key()`,
-    # but that requires a newer minimum version of cryptography.
-    return req.is_signature_valid and req.public_key().public_bytes(
-        serialization.Encoding.DER, serialization.PublicFormat.SubjectPublicKeyInfo
-    ) == pkey.public_key().public_bytes(
-        serialization.Encoding.DER, serialization.PublicFormat.SubjectPublicKeyInfo
-    )
+    return req.is_signature_valid and req.public_key() == pkey.public_key()
 
 
-def import_csr_file(csrfile: str, data: bytes) -> Tuple[int, util.CSR, List[str]]:
+def import_csr_file(
+    csrfile: str, data: bytes
+) -> Tuple[acme_crypto_util.Format, util.CSR, List[str]]:
     """Import a CSR file, which can be either PEM or DER.
 
     :param str csrfile: CSR filename
     :param bytes data: contents of the CSR file
 
-    :returns: (`crypto.FILETYPE_PEM`,
+    :returns: (`acme_crypto_util.Format.PEM`,
                util.CSR object representing the CSR,
                list of domains requested in the CSR)
     :rtype: tuple
 
     """
-    PEM = crypto.FILETYPE_PEM
-    load = crypto.load_certificate_request
     try:
         # Try to parse as DER first, then fall back to PEM.
-        csr = load(crypto.FILETYPE_ASN1, data)
-    except crypto.Error:
+        csr = x509.load_der_x509_csr(data)
+    except ValueError:
         try:
-            csr = load(PEM, data)
-        except crypto.Error:
+            csr = x509.load_pem_x509_csr(data)
+        except ValueError:
             raise errors.Error("Failed to parse CSR file: {0}".format(csrfile))
 
-    domains = _get_names_from_loaded_cert_or_req(csr)
+    domains = acme_crypto_util.get_names_from_subject_and_extensions(csr.subject, csr.extensions)
     # Internally we always use PEM, so re-encode as PEM before returning.
-    data_pem = crypto.dump_certificate_request(PEM, csr)
-    return PEM, util.CSR(file=csrfile, data=data_pem, form="pem"), domains
+    data_pem = csr.public_bytes(serialization.Encoding.PEM)
+    return (
+        acme_crypto_util.Format.PEM,
+        util.CSR(file=csrfile, data=data_pem, form="pem"),
+        domains,
+    )
 
 
 def make_key(bits: int = 2048, key_type: str = "rsa",
@@ -220,7 +216,8 @@ def make_key(bits: int = 2048, key_type: str = "rsa",
 
     :returns: new RSA or ECDSA key in PEM form with specified number of bits
               or of type ec_curve when key_type ecdsa is used.
-    :rtype: str
+    :rtype: bytes
+
     """
     key: Union[rsa.RSAPrivateKey, ec.EllipticCurvePrivateKey]
     if key_type == 'rsa':
@@ -251,7 +248,7 @@ def make_key(bits: int = 2048, key_type: str = "rsa",
         raise errors.Error("Invalid key_type specified: {}.  Use [rsa|ecdsa]".format(key_type))
     return key.private_bytes(
         encoding=Encoding.PEM,
-        format=PrivateFormat.TraditionalOpenSSL,
+        format=PrivateFormat.PKCS8,
         encryption_algorithm=NoEncryption()
     )
 
@@ -265,11 +262,14 @@ def valid_privkey(privkey: Union[str, bytes]) -> bool:
     :rtype: bool
 
     """
+    if isinstance(privkey, str):
+        privkey = privkey.encode()
     try:
-        return crypto.load_privatekey(
-            crypto.FILETYPE_PEM, privkey).check()
-    except (TypeError, crypto.Error):
+        serialization.load_pem_private_key(privkey, password=None)
+    except ValueError:
         return False
+    else:
+        return True
 
 
 def verify_renewable_cert(renewable_cert: interfaces.RenewableCert) -> None:
@@ -408,82 +408,84 @@ def pyopenssl_load_certificate(data: bytes) -> Tuple[crypto.X509, int]:
         str(error) for error in openssl_errors)))
 
 
-def _load_cert_or_req(cert_or_req_str: bytes,
-                      load_func: Callable[[int, bytes], Union[crypto.X509, crypto.X509Req]],
-                      typ: int = crypto.FILETYPE_PEM) -> Union[crypto.X509, crypto.X509Req]:
-    try:
-        return load_func(typ, cert_or_req_str)
-    except crypto.Error as err:
-        logger.debug("", exc_info=True)
-        logger.error("Encountered error while loading certificate or csr: %s", str(err))
-        raise
-
-
-def _get_sans_from_cert_or_req(cert_or_req_str: bytes,
-                               load_func: Callable[[int, bytes], Union[crypto.X509,
-                                                                       crypto.X509Req]],
-                               typ: int = crypto.FILETYPE_PEM) -> List[str]:
-    # pylint: disable=protected-access
-    return acme_crypto_util._pyopenssl_cert_or_req_san(_load_cert_or_req(
-        cert_or_req_str, load_func, typ))
-
-
-def get_sans_from_cert(cert: bytes, typ: int = crypto.FILETYPE_PEM) -> List[str]:
+def get_sans_from_cert(
+    cert: bytes, typ: Union[acme_crypto_util.Format, int] = acme_crypto_util.Format.PEM
+) -> List[str]:
     """Get a list of Subject Alternative Names from a certificate.
 
     :param str cert: Certificate (encoded).
-    :param typ: `crypto.FILETYPE_PEM` or `crypto.FILETYPE_ASN1`
+    :param Format typ: Which format the `cert` bytes are in.
 
     :returns: A list of Subject Alternative Names.
     :rtype: list
 
     """
-    return _get_sans_from_cert_or_req(
-        cert, crypto.load_certificate, typ)
+    typ = acme_crypto_util.Format(typ)
+    if typ == acme_crypto_util.Format.PEM:
+        x509_cert = x509.load_pem_x509_certificate(cert)
+    else:
+        assert typ == acme_crypto_util.Format.DER
+        x509_cert = x509.load_der_x509_certificate(cert)
+
+    try:
+        san_ext = x509_cert.extensions.get_extension_for_class(
+            x509.SubjectAlternativeName
+        )
+    except x509.ExtensionNotFound:
+        return []
+
+    return san_ext.value.get_values_for_type(x509.DNSName)
 
 
-def _get_names_from_cert_or_req(cert_or_req: bytes,
-                                load_func: Callable[[int, bytes], Union[crypto.X509,
-                                                                        crypto.X509Req]],
-                                typ: int) -> List[str]:
-    loaded_cert_or_req = _load_cert_or_req(cert_or_req, load_func, typ)
-    return _get_names_from_loaded_cert_or_req(loaded_cert_or_req)
-
-
-def _get_names_from_loaded_cert_or_req(loaded_cert_or_req: Union[crypto.X509, crypto.X509Req]
-                                       ) -> List[str]:
-    # pylint: disable=protected-access
-    return acme_crypto_util._pyopenssl_cert_or_req_all_names(loaded_cert_or_req)
-
-
-def get_names_from_cert(cert: bytes, typ: int = crypto.FILETYPE_PEM) -> List[str]:
+def get_names_from_cert(
+    cert: bytes, typ: Union[acme_crypto_util.Format, int] = acme_crypto_util.Format.PEM
+) -> List[str]:
     """Get a list of domains from a cert, including the CN if it is set.
 
     :param str cert: Certificate (encoded).
-    :param typ: `crypto.FILETYPE_PEM` or `crypto.FILETYPE_ASN1`
+    :param Format typ: Which format the `cert` bytes are in.
 
     :returns: A list of domain names.
     :rtype: list
 
     """
-    return _get_names_from_cert_or_req(
-        cert, crypto.load_certificate, typ)
+    typ = acme_crypto_util.Format(typ)
+    if typ == acme_crypto_util.Format.PEM:
+        x509_cert = x509.load_pem_x509_certificate(cert)
+    else:
+        assert typ == acme_crypto_util.Format.DER
+        x509_cert = x509.load_der_x509_certificate(cert)
+    return acme_crypto_util.get_names_from_subject_and_extensions(
+        x509_cert.subject, x509_cert.extensions
+    )
 
 
-def get_names_from_req(csr: bytes, typ: int = crypto.FILETYPE_PEM) -> List[str]:
+def get_names_from_req(
+    csr: bytes, typ: Union[acme_crypto_util.Format, int] = acme_crypto_util.Format.PEM
+) -> List[str]:
     """Get a list of domains from a CSR, including the CN if it is set.
 
     :param str csr: CSR (encoded).
-    :param typ: `crypto.FILETYPE_PEM` or `crypto.FILETYPE_ASN1`
+    :param acme_crypto_util.Format typ: Which format the `csr` bytes are in.
     :returns: A list of domain names.
     :rtype: list
 
     """
-    return _get_names_from_cert_or_req(csr, crypto.load_certificate_request, typ)
+    typ = acme_crypto_util.Format(typ)
+    if typ == acme_crypto_util.Format.PEM:
+        x509_req = x509.load_pem_x509_csr(csr)
+    else:
+        assert typ == acme_crypto_util.Format.DER
+        x509_req = x509.load_der_x509_csr(csr)
+    return acme_crypto_util.get_names_from_subject_and_extensions(
+        x509_req.subject, x509_req.extensions
+    )
 
 
-def dump_pyopenssl_chain(chain: Union[List[crypto.X509], List[josepy.ComparableX509]],
-                         filetype: int = crypto.FILETYPE_PEM) -> bytes:
+def dump_pyopenssl_chain(
+    chain: Union[List[crypto.X509], List[josepy.ComparableX509]],
+    filetype: Union[acme_crypto_util.Format, int] = acme_crypto_util.Format.PEM,
+) -> bytes:
     """Dump certificate chain into a bundle.
 
     :param list chain: List of `crypto.X509` (or wrapped in
@@ -504,7 +506,9 @@ def notBefore(cert_path: str) -> datetime.datetime:
     :rtype: :class:`datetime.datetime`
 
     """
-    return _notAfterBefore(cert_path, crypto.X509.get_notBefore)
+    with open(cert_path, "rb") as f:
+        cert = x509.load_pem_x509_certificate(f.read())
+    return cert.not_valid_before_utc
 
 
 def notAfter(cert_path: str) -> datetime.datetime:
@@ -516,35 +520,9 @@ def notAfter(cert_path: str) -> datetime.datetime:
     :rtype: :class:`datetime.datetime`
 
     """
-    return _notAfterBefore(cert_path, crypto.X509.get_notAfter)
-
-
-def _notAfterBefore(cert_path: str,
-                    method: Callable[[crypto.X509], Optional[bytes]]) -> datetime.datetime:
-    """Internal helper function for finding notbefore/notafter.
-
-    :param str cert_path: path to a cert in PEM format
-    :param function method: one of ``crypto.X509.get_notBefore``
-        or ``crypto.X509.get_notAfter``
-
-    :returns: the notBefore or notAfter value from the cert at cert_path
-    :rtype: :class:`datetime.datetime`
-
-    """
-    # pylint: disable=redefined-outer-name
     with open(cert_path, "rb") as f:
-        x509 = crypto.load_certificate(crypto.FILETYPE_PEM, f.read())
-    # pyopenssl always returns bytes
-    timestamp = method(x509)
-    if not timestamp:
-        raise errors.Error("Error while invoking timestamp method, None has been returned.")
-    reformatted_timestamp = [timestamp[0:4], b"-", timestamp[4:6], b"-",
-                             timestamp[6:8], b"T", timestamp[8:10], b":",
-                             timestamp[10:12], b":", timestamp[12:]]
-    # pyrfc3339 always uses the type `str`
-    timestamp_bytes = b"".join(reformatted_timestamp)
-    timestamp_str = timestamp_bytes.decode('ascii')
-    return pyrfc3339.parse(timestamp_str)
+        cert = x509.load_pem_x509_certificate(f.read())
+    return cert.not_valid_after_utc
 
 
 def sha256sum(filename: str) -> str:
@@ -613,10 +591,9 @@ def get_serial_from_cert(cert_path: str) -> int:
     :returns: serial number of the certificate
     :rtype: int
     """
-    # pylint: disable=redefined-outer-name
     with open(cert_path, "rb") as f:
-        x509 = crypto.load_certificate(crypto.FILETYPE_PEM, f.read())
-    return x509.get_serial_number()
+        cert = x509.load_pem_x509_certificate(f.read())
+    return cert.serial_number
 
 
 def find_chain_with_issuer(fullchains: List[str], issuer_cn: str,